requirements.txt
gradio>=4.0.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
numpy>=1.21.0